Module google_deploymentmanager2::api

Structs

AuditConfig
Specifies the audit configuration for a service. The configuration determines which permission types are logged, and what identities, if any, are exempted from logging. An AuditConfig must have one or more AuditLogConfigs. If there are AuditConfigs for both allServices and a specific service, the union of the two AuditConfigs is used for that service: the log_types specified in each AuditConfig are enabled, and the exempted_members in each AuditLogConfig are exempted. Example Policy with multiple AuditConfigs: { "audit_configs": [ { "service": "allServices", "audit_log_configs": [ { "log_type": "DATA_READ", "exempted_members": [ "user:jose@example.com" ] }, { "log_type": "DATA_WRITE" }, { "log_type": "ADMIN_READ" } ] }, { "service": "sampleservice.googleapis.com", "audit_log_configs": [ { "log_type": "DATA_READ" }, { "log_type": "DATA_WRITE", "exempted_members": [ "user:aliya@example.com" ] } ] } ] } For sampleservice, this policy enables DATA_READ, DATA_WRITE and ADMIN_READ logging. It also exempts jose@example.com from DATA_READ logging, and aliya@example.com from DATA_WRITE logging.

AuditLogConfig
Provides the configuration for logging a type of permissions. Example: { "audit_log_configs": [ { "log_type": "DATA_READ", "exempted_members": [ "user:jose@example.com" ] }, { "log_type": "DATA_WRITE" } ] } This enables 'DATA_READ' and 'DATA_WRITE' logging, while exempting jose@example.com from DATA_READ logging.

Binding
Associates members with a role.

ConfigFile
There is no detailed description.

Deployment
There is no detailed description.

DeploymentCancelPreviewCall
Cancels and removes the preview currently associated with the deployment.

DeploymentDeleteCall
Deletes a deployment and all of the resources in the deployment.

DeploymentGetCall
Gets information about a specific deployment.

DeploymentGetIamPolicyCall
Gets the access control policy for a resource. May be empty if no such policy or resource exists.

DeploymentInsertCall
Creates a deployment and all of the resources described by the deployment manifest.

DeploymentLabelEntry
Label object for Deployments

DeploymentListCall
Lists all deployments for a given project.

DeploymentManager
Central instance to access all DeploymentManager related resource activities

DeploymentMethods
A builder providing access to all methods supported on deployment resources. It is not used directly, but through the DeploymentManager hub.

DeploymentPatchCall
Patches a deployment and all of the resources described by the deployment manifest.

DeploymentSetIamPolicyCall
Sets the access control policy on the specified resource. Replaces any existing policy.

DeploymentStopCall
Stops an ongoing operation. This does not roll back any work that has already been completed, but prevents any new work from being started.

DeploymentTestIamPermissionCall
Returns permissions that a caller has on the specified resource.

DeploymentUpdate
There is no detailed description.

DeploymentUpdateCall
Updates a deployment and all of the resources described by the deployment manifest.

DeploymentUpdateLabelEntry
Label object for DeploymentUpdate

DeploymentsCancelPreviewRequest
There is no detailed description.

DeploymentsListResponse
A response containing a partial list of deployments and a page token used to build the next request if the request has been truncated.

DeploymentsStopRequest
There is no detailed description.

Expr
Represents a textual expression in the Common Expression Language (CEL) syntax. CEL is a C-like expression language. The syntax and semantics of CEL are documented at https://github.com/google/cel-spec. Example (Comparison): title: "Summary size limit" description: "Determines if a summary is less than 100 chars" expression: "document.summary.size() < 100" Example (Equality): title: "Requestor is owner" description: "Determines if requestor is the document owner" expression: "document.owner == request.auth.claims.email" Example (Logic): title: "Public documents" description: "Determine whether the document should be publicly visible" expression: "document.type != 'private' && document.type != 'internal'" Example (Data Manipulation): title: "Notification string" description: "Create a notification string with a timestamp." expression: "'New message received at ' + string(document.create_time)" The exact variables and functions that may be referenced within an expression are determined by the service that evaluates it. See the service documentation for additional information.

GlobalSetPolicyRequest
There is no detailed description.

ImportFile
There is no detailed description.

Manifest
There is no detailed description.

ManifestGetCall
Gets information about a specific manifest.

ManifestListCall
Lists all manifests for a given deployment.

ManifestMethods
A builder providing access to all methods supported on manifest resources. It is not used directly, but through the DeploymentManager hub.

ManifestsListResponse
A response containing a partial list of manifests and a page token used to build the next request if the request has been truncated.

Operation
Represents an Operation resource. Google Compute Engine has three Operation resources: * Global * Regional * Zonal You can use an operation resource to manage asynchronous API requests. For more information, read Handling API responses. Operations can be global, regional or zonal. - For global operations, use the globalOperations resource. - For regional operations, use the regionOperations resource. - For zonal operations, use the zonalOperations resource. For more information, read Global, Regional, and Zonal Resources.

OperationError
[Output Only] If errors are generated during processing of the operation, this field will be populated.

OperationErrorErrors
[Output Only] The array of errors encountered while processing this operation.

OperationGetCall
Gets information about a specific operation.

OperationListCall
Lists all operations for a project.

OperationMethods
A builder providing access to all methods supported on operation resources. It is not used directly, but through the DeploymentManager hub.

OperationWarnings
[Output Only] If warning messages are generated during processing of the operation, this field will be populated.

OperationWarningsData
[Output Only] Metadata about this warning in key: value format. For example: "data": [ { "key": "scope", "value": "zones/us-east1-d" }

OperationsListResponse
A response containing a partial list of operations and a page token used to build the next request if the request has been truncated.

Policy
An Identity and Access Management (IAM) policy, which specifies access controls for Google Cloud resources. A Policy is a collection of bindings. A binding binds one or more members to a single role. Members can be user accounts, service accounts, Google groups, and domains (such as G Suite). A role is a named list of permissions; each role can be an IAM predefined role or a user-created custom role. For some types of Google Cloud resources, a binding can also specify a condition, which is a logical expression that allows access to a resource only if the expression evaluates to true. A condition can add constraints based on attributes of the request, the resource, or both. To learn which resources support conditions in their IAM policies, see the IAM documentation. JSON example: { "bindings": [ { "role": "roles/resourcemanager.organizationAdmin", "members": [ "user:mike@example.com", "group:admins@example.com", "domain:google.com", "serviceAccount:my-project-id@appspot.gserviceaccount.com" ] }, { "role": "roles/resourcemanager.organizationViewer", "members": [ "user:eve@example.com" ], "condition": { "title": "expirable access", "description": "Does not grant access after Sep 2020", "expression": "request.time < timestamp('2020-10-01T00:00:00.000Z')", } } ], "etag": "BwWWja0YfJA=", "version": 3 } YAML example: bindings: - members: - user:mike@example.com - group:admins@example.com - domain:google.com - serviceAccount:my-project-id@appspot.gserviceaccount.com role: roles/resourcemanager.organizationAdmin - members: - user:eve@example.com role: roles/resourcemanager.organizationViewer condition: title: expirable access description: Does not grant access after Sep 2020 expression: request.time < timestamp('2020-10-01T00:00:00.000Z') - etag: BwWWja0YfJA= - version: 3 For a description of IAM and its features, see the IAM documentation.

Resource
There is no detailed description.

ResourceAccessControl
The access controls set on the resource.

ResourceGetCall
Gets information about a single resource.

ResourceListCall
Lists all resources in a given deployment.

ResourceMethods
A builder providing access to all methods supported on resource resources. It is not used directly, but through the DeploymentManager hub.

ResourceUpdate
There is no detailed description.

ResourceUpdateError
Output only. If errors are generated during update of the resource, this field will be populated.

ResourceUpdateErrorErrors
[Output Only] The array of errors encountered while processing this operation.

ResourceUpdateWarnings
Output only. If warning messages are generated during processing of this resource, this field will be populated.

ResourceUpdateWarningsData
[Output Only] Metadata about this warning in key: value format. For example: "data": [ { "key": "scope", "value": "zones/us-east1-d" }

ResourceWarnings
Output only. If warning messages are generated during processing of this resource, this field will be populated.

ResourceWarningsData
[Output Only] Metadata about this warning in key: value format. For example: "data": [ { "key": "scope", "value": "zones/us-east1-d" }

ResourcesListResponse
A response containing a partial list of resources and a page token used to build the next request if the request has been truncated.

TargetConfiguration
There is no detailed description.

TestPermissionsRequest
There is no detailed description.

TestPermissionsResponse
There is no detailed description.

Type
A resource type supported by Deployment Manager.

TypeListCall
Lists all resource types for Deployment Manager.

TypeMethods
A builder providing access to all methods supported on type resources. It is not used directly, but through the DeploymentManager hub.

TypesListResponse
A response that returns all Types supported by Deployment Manager

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.