Struct google_deploymentmanager2::DeploymentMethods

pub struct DeploymentMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on deployment resources. It is not used directly, but through the DeploymentManager hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_deploymentmanager2 as deploymentmanager2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use deploymentmanager2::DeploymentManager;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = DeploymentManager::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `cancel_preview(...)`, `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)`, `stop(...)` and `update(...)`
// to build up your call.
let rb = hub.deployments();

Methods

impl<'a, C, A> DeploymentMethods<'a, C, A>

fn update(&self, request: Deployment, project: &str, deployment: &str) -> DeploymentUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a deployment and all of the resources described by the deployment manifest.

Arguments

request - No description provided.

project - The project ID for this request.

deployment - The name of the deployment for this request.

fn get(&self, project: &str, deployment: &str) -> DeploymentGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets information about a specific deployment.

Arguments

project - The project ID for this request.

deployment - The name of the deployment for this request.

fn insert(&self, request: Deployment, project: &str) -> DeploymentInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a deployment and all of the resources described by the deployment manifest.

Arguments

request - No description provided.

project - The project ID for this request.

fn cancel_preview(&self, request: DeploymentsCancelPreviewRequest, project: &str, deployment: &str) -> DeploymentCancelPreviewCall<'a, C, A>

Create a builder to help you perform the following task:

Cancels and removes the preview currently associated with the deployment.

Arguments

request - No description provided.

project - The project ID for this request.

deployment - The name of the deployment for this request.

fn list(&self, project: &str) -> DeploymentListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all deployments for a given project.

Arguments

project - The project ID for this request.

fn patch(&self, request: Deployment, project: &str, deployment: &str) -> DeploymentPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a deployment and all of the resources described by the deployment manifest. This method supports patch semantics.

Arguments

request - No description provided.

project - The project ID for this request.

deployment - The name of the deployment for this request.

fn stop(&self, request: DeploymentsStopRequest, project: &str, deployment: &str) -> DeploymentStopCall<'a, C, A>

Create a builder to help you perform the following task:

Stops an ongoing operation. This does not roll back any work that has already been completed, but prevents any new work from being started.

Arguments

request - No description provided.

project - The project ID for this request.

deployment - The name of the deployment for this request.

fn delete(&self, project: &str, deployment: &str) -> DeploymentDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a deployment and all of the resources in the deployment.

Arguments

project - The project ID for this request.

deployment - The name of the deployment for this request.