Struct google_deploymentmanager2::DeploymentUpdate

pub struct DeploymentUpdate {
    pub labels: Option<Vec<DeploymentUpdateLabelEntry>>,
    pub manifest: Option<String>,
}

This type is not used in any activity, and only used as part of another schema.

Fields
labels: Option<Vec<DeploymentUpdateLabelEntry>>

[Output Only] Map of labels; provided by the client when the resource is created or updated. Specifically: Label keys must be between 1 and 63 characters long and must conform to the following regular expression: a-z? Label values must be between 0 and 63 characters long and must conform to the regular expression (a-z?)?

manifest: Option<String>

[Output Only] URL of the manifest representing the update configuration of this deployment.