Struct google_deploymentmanager2::ResourceType

pub struct ResourceType {
    pub final_properties: Option<String>,
    pub update_time: Option<String>,
    pub name: Option<String>,
    pub warnings: Option<Vec<ResourceWarnings>>,
    pub url: Option<String>,
    pub update: Option<ResourceUpdate>,
    pub manifest: Option<String>,
    pub properties: Option<String>,
    pub insert_time: Option<String>,
    pub type_: Option<String>,
    pub id: Option<String>,
}

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get resources (response)

list resources (none)

Fields

final_properties

[Output Only] The evaluated properties of the resource with references expanded. Returned as serialized YAML.

update_time

[Output Only] Timestamp when the resource was updated, in RFC3339 text format .

name

[Output Only] The name of the resource as it appears in the YAML config.

warnings

[Output Only] If warning messages are generated during processing of this resource, this field will be populated.

url

[Output Only] The URL of the actual resource.

update

[Output Only] If Deployment Manager is currently updating or previewing an update to this resource, the updated configuration appears here.

manifest

[Output Only] URL of the manifest representing the current configuration of this resource.

properties

[Output Only] The current properties of the resource before any references have been filled in. Returned as serialized YAML.

insert_time

[Output Only] Timestamp when the resource was created or acquired, in RFC3339 text format .

type_

[Output Only] The type of the resource, for example compute.v1.instance, or replicaPools.v1beta2.instanceGroupManager.

id

[Output Only] Unique identifier for the resource; defined by the server.